requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-multipart==0.0.9
pillow==10.*
numpy==1.*
httpx==0.27.*      # needed by fastapi TestClient via starlette
pytest==8.*
prometheus-fastapi-instrumentator==7.0.0
python-dotenv==1.0.*
requests==2.32.*
streamlit==1.38.*